// src/_shared/enums/culture.enums.ts
export enum CultureEnums {
  aszisz = 'aszisz',
  crantai = 'crantai',
  dwoon = 'dwoon',
  dzsad = 'dzsad',
  elf = 'elf',
  enoszukei = 'enoszukei',
  erv = 'erv',
  goblin = 'goblin',
  gorviki = 'gorviki',
  ilanori = 'ilanori',
  korg = 'korg',
  krani = 'krani',
  kyr_toroni = 'kyr_toroni',
  nomad = 'nomad',
  ork = 'ork',
  pyarroni = 'pyarroni',
  shadoni = 'shadoni',
  torpe = 'torpe',
}